requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=1.10.0
aiofiles>=0.8.0
aiothrift>=0.1.0
asyncio>=3.4.3
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5
plotly>=5.13.0
iotdb-client>=2.0.8
pydantic-settings>=2.0.0